requirements.txt
openai
requests
anthropic
python-dotenv
Pillow
scikit-learn
pandas
numpy
matplotlib